Convert mesh_adv feature to bindless, other additions/changes necessary to support it (#227)
 - Add queries to find descriptors/push constants on rafx_api::RafxRootSignature
 - Push constant support for metal/vulkan
 - Allow buffer barriers to apply to a specific range of the buffer
 - Allow asset plugins to tick while doing a blocking load
 - Refactor upload queue to reduce code duplication/support more usage scenarios
 - Support depth bias in materials
 - Improvements to the shader processor
 - Replace apply_setup(), revert_set() with begin_submit_node_batch()
 - on_frame_complete(), free() callback for asset types

+++ rafx-api/src/backends/metal/pipeline.rs
@@ -188,8 +188,7 @@ impl RafxPipelineMetal {
         let mtl_triangle_fill_mode = pipeline_def.rasterizer_state.fill_mode.into();
         let mtl_front_facing_winding = pipeline_def.rasterizer_state.front_face.into();
         let mtl_depth_bias = pipeline_def.rasterizer_state.depth_bias as f32;
-        let mtl_depth_bias_slope_scaled =
-            pipeline_def.rasterizer_state.depth_bias_slope_scaled as f32;
+        let mtl_depth_bias_slope_scaled = pipeline_def.rasterizer_state.depth_bias_slope_scaled;
         let mtl_depth_clip_mode = if pipeline_def.rasterizer_state.depth_clamp_enable {
             metal_rs::MTLDepthClipMode::Clamp
         } else {
